+++ src/rv64core/mmu/sv48.rs
@@ -69,7 +69,7 @@ impl PAops for Sv48PA {
         }
     }
     fn offset(&self) -> usize {
-        self.offset() as usize
+        self.offset()
     }
 
     fn set_offset(&mut self, val: usize) {
